connectfour v0.1.0 Board

Documentation for ConnectFour.

Functions

default()

Returns an empty board with the default columns and rows.

Examples

iex> Board.default[:rows]
6
iex> Board.default[:cols]
7
iex> Board.default[:indicies][{0, 0}]
:empty
iex> Board.default |> Board.is_empty?
true

directions()

Returns a list containing indicies representing compass directions. Used to explore a board from a starting position.

Examples

iex> Board.directions
[ {-1, -1}, {-1, 0}, {-1, 1}, {0, -1}, {0, 1}, {1, -1}, {1, 0}, {1, 1} ]

get(board, index)

Gets the marker at the board index

Examples

iex> Board.default |> Board.set({0, 0}, :m) |> Board.get({0, 0}) == :m
true

indicies(board)

Retrieves all of the indicies, like {0, 0}, of the board as a list.

is_empty?(board)

Returns true if the passed board is empty.

Examples

iex> Board.default |> Board.is_empty?
true

is_empty_at?(board, index)

Returns true if the board is empty at the passed index.

Examples

iex> Board.is_empty_at?(Board.default, {0, 0})
true

is_full?(board)

Returns true if there are no empty spots on the board.

iex> Board.default |> Board.indicies
...> |> Enum.reduce(Board.default,
...>                fn(ix, b) -> b |> Board.set(ix, :m) end)
...> |> Board.is_full?
true

marker_str(arg1)

Converts the passed symbol into a string for display.

render(board)

Renders the board as a string.

Examples

iex> Board.default |> Board.set({0, 0}, :x) |> Board.render
".......
.......
.......
.......
.......
X......"

set(board, index, marker)

Sets the marker at the board index

Examples

iex> Board.default |> Board.set({0, 0}, :m) |> Board.is_empty_at?({0, 0})
false