connectfour v0.1.0 Game

Documentation for ConnectFour.

Functions

check_board_pos(board, index, marker, direction, steps)

Returns true if the board contains a matching marker at the passed index and direction, for the passed number of steps. Is used to find adjacent indicies with the same marker.

default()

A default game with an empty board

Examples

iex> Game.default[:current_marker]
:x

first_empty_row(board, column)

Returns the first empty row in the passed column.

Examples

iex> Game.default[:board] |> Game.first_empty_row(0)
0
iex> Game.default |> Game.make_move(0)
...> |> Map.get(:board) |> Game.first_empty_row(0)
1

index_in_direction(direction, index)

Use to get the adjusted index in the direction relative to the passed index

iex> Game.index_in_direction({1, 0}, {0, 0})
{1, 0}

is_win?(game)

Determines if any marker has won.

make_move(game, column)

Makes a move by setting the marker in the first empty row of the column.

Examples

iex> Game.default |> Game.make_move(0)
...> |> Map.get(:board) |> Board.get({0, 0})
:x

next_marker(game)

Returns the next marker to play, based on the current marker.

iex> Game.default |> Game.next_marker
:o
iex> Game.default |> Game.make_move(0) |> Game.next_marker
:x

render(game)

Renders the game as a string.

win_at_index?(game, index)

Determines if the marker at the passed index has won the game, by comparing adjacent markers in the compass directions.